requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cors==1.0.1
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
filelock==3.0.12
future==0.18.2
gevent==21.8.0
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
tldextract==3.1.0
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
zope.event==4.5.0
zope.interface==5.4.0
